Add limelight override settings to smartdashboard
Right now we aren't able to use the limelight during the day.
Eventually this is going to block us from being able to tune
the shooter with respect to distance-from-target. Rather than
being forced to tune the shooter after dark, we can instead
use smartdashbaord to inject feasible limelight observations.